third_party/blink/web_tests/fast/dom/Element/dimension-properties-unrendered-expected.txt

This test checks the scroll, client and offset dimension properties return "0" instead of undefined for an unrendered element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS testDiv.offsetLeft is 0
PASS testDiv.offsetTop is 0
PASS testDiv.offsetWidth is 0
PASS testDiv.offsetHeight is 0
PASS testDiv.clientWidth is 0
PASS testDiv.clientHeight is 0
PASS testDiv.scrollLeft is 0
PASS testDiv.scrollTop is 0
PASS testDiv.scrollWidth is 0
PASS testDiv.scrollHeight is 0
PASS successfullyParsed is true

TEST COMPLETE
